bear-requirements.txt: Use PipRequirement in bears
Remove all bear requirements from requirements.txt,
and use .ci/generate_bear_requirements.py to create
and verify that the bear-requirements is identical to
the PipRequirement found in the bears.

Updates setup.py to add an extra 'alldeps' that is be used
by CI to transition to explicitly requesting additional
dependencies so 100% coverage and `coala --ci` work correctly.