Gem requirements: Use semantic version ranges
Without any version operator, gem defaults to `>=`.
Our npm and pip dependencies use semantic versioning
where upgrades are only allowed according to the
level specified in the version string.

In ruby gem version specs, this is called pessimistic versions
and is prefixed with the operator `~>`.

generate_bear_requirements.py is modified to use `~>` by default,
so it is inline with our pip and npm version specifications,
and it is easy to constrain ruby gem versions when breakages
occur.  It also means that `bear-requirements.yaml` always
contains an explicit version operator whenever there is a
version, which reduces misunderstanding given that each
platform manager has different default operators when
only a version is given.

Related to https://github.com/coala/coala-bears/issues/2558