Revert ".gitignore: Remove package-lock.json"
This reverts commit c77ee658178d4aaac7b66a63ffd87c57749a398e.

coala bears does not control the installed versions on users
machines.
The package.json in this repository is generated to aide installing,
and is not used to build a product and is therefore not locked!
CI fails when the lock is not maintained, and we have no reason
to maintain a lock file.

Related to https://github.com/coala/coala-bears/issues/2508