CPDBear: Detect appropriate linter executable
Describe current executable requirements of cpd or run.sh,
replacing check_prerequisites with Requirement classes.

Also remove explicit invocation using `bash`, which was
ineffective as which('run.sh') would only be successful
if the script was executable, and the run.sh script is not
written to be compatible with invocation on Windows.

This de-supports any usage where `pmd` existed and was a script
which could invoke `cpd`.  It fixes usage where `pmd` is not a
script, like it is an executable shim when installed with choco.

Also remove duplication of `CPDBear` in generate_bear_metadata.py
which was intended to make 'cpd' deselectable, but was broken.
It is now mostly unnecessary as the ExecutableRequirement classes
provide the `cpd` tag, however the `java` tag is still manual.

Reduce the required coverage percentage to 98%.

Fixes https://github.com/coala/coala-bears/issues/2937
Related to https://github.com/coala/coala-bears/issues/2908